requirements.txt
Django
django-location-field==2.0.0b0
